new search
I moved the add make and model forms to the make and model page. The
add scooter page is moved to the all scooters page. The member
mfilter.php now filters members by date and the msfilter.php page does
what mfilter used to do only now it does it with a search input.